// ejerciciosPruebas/copy/ej2/angularFormulario/src/app/formularios-anidados/interfaces/user.interface.ts
export interface User {
  name: string;
  age: number;
  email: string;
  user: string;
  pass: string;
  repeatPass: string;
}

export interface User2 {
  name: string;
  age: number;
  email: string;
  user: string;
  pass: string;
  repeatPass: string;
  hobbiesForm: string[];
}

export interface address {
  address: string;
  city: string;
  code: number;
}